[quant][core] Added support for quantizing a constant value and broadcasting the assignment to an entire tensor
Summary:
Support in QuantizedOpKernels.cpp exists for quantizing a
fp tensor and copying the quantized tensor to the output (quantized)
tensor. This PR adds analogous support for quantizing a scalar input and
broadcasting the quantized value to the entire tensor.

This PR is intended to be used for fill_quantized, where, currently, a
tensor filled with the same quantized value needs to be created in order
to be assigned to an output quantized tensor, resulting in an
inefficient fill operation.

Test plan:

ghstack-source-id: c46d1fa84bdb37a7428de692b561d3e068911b89
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78562